requirements.txt
streamlit
pandas
numpy
plotly
matplotlib
seaborn
scikit-learn
openpyxl
xgboost
python-pptx
Pillow
shap
joblib
